[Ticket:32775] There was an unhandled exception occurring when deserializing the Json response from Hubspot for one of the contacts. The value was bigger than the datatype specified for it (int) so it was updated to long.
git-svn-id: https://teamsupport.svn.beanstalkapp.com/main/trunk@4716 02885c34-e171-44ab-8435-42fb7a3c3aaf

+++ TeamSupport.ServiceLibrary/HubSpotSources/Objects/Contacts.cs
@@ -80,7 +80,7 @@ namespace TeamSupport.ServiceLibrary.HubSpotSources.Objects
       [DataMember(Name = "contacts")]
       public List<Contact> contacts { get; set; }
       [DataMember(Name = "time-offset")]
-      public int timeOffset { get; set; }
+      public long timeOffset { get; set; }
       [DataMember(Name = "query")]
       public string query { get; set; }
       [DataMember(Name = "vid-offset")]
